requirements.txt
streamlit
requests
pydantic
llama-index
llama-index-vector-stores-chroma
chromadb
python-dotenv
beautifulsoup4
llama-index-embeddings-huggingface
lxml